+++ src/ExtendableFEM.jl
@@ -66,7 +66,7 @@ using LinearAlgebra: LinearAlgebra, copyto!, isposdef, mul!, norm
 using LinearSolve: LinearSolve, LinearProblem, UMFPACKFactorization, deleteat!,
     init, solve
 using Printf: Printf, @printf
-using SparseArrays: SparseArrays, AbstractSparseArray, SparseMatrixCSC, nnz,
+using SparseArrays: SparseArrays, AbstractSparseArray, SparseMatrixCSC, findnz, nnz,
     nzrange, rowvals, sparse
 using SparseDiffTools: SparseDiffTools, ForwardColorJacCache,
     forwarddiff_color_jacobian!, matrix_colors
@@ -129,6 +129,7 @@ export replace_operator!
 
 include("helper_functions.jl")
 export get_periodic_coupling_info
+export get_periodic_coupling_matrix
 export tmul!
 
 include("tensors.jl")